refactor: 도메인 디렉토리 분리
