+++ src/test/java/christmas/domain/plan/BadgeTest.java
@@ -1,12 +1,10 @@
-package christmas.domain;
+package christmas.domain.plan;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class BadgeTest {
