+++ src/test/java/christmas/domain/visitor/MenuOrderTest.java
@@ -1,5 +1,6 @@
-package christmas.domain;
+package christmas.domain.visitor;
 
+import christmas.domain.visitor.MenuOrder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
